% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{access_client}
\alias{access_client}
\title{Access client variables}
\usage{
access_client()
}
\description{
Retrieves the relevant client variables
}
\details{
A helper function that retrieves the API details set as environment variables
by \code{\link{setup_client()}}.
}
