% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.account}
\alias{as.data.frame.account}
\title{Convert account class to data.frame}
\usage{
\method{as.data.frame}{account}(x, ...)
}
\arguments{
\item{x}{The account class returned by \code{\link{whoami()}}}
}
\value{
A data.frame containing the user's account information
}
\description{
Convert account class to data.frame
}
