% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.calibration_models}
\alias{as.data.frame.calibration_models}
\title{Coerce calibration_models class to data.frame}
\usage{
\method{as.data.frame}{calibration_models}(x, ...)
}
\description{
Coerce calibration_models class to data.frame
}
