% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.device_data}
\alias{as.data.frame.device_data}
\title{Coerce device data to a data.frame}
\usage{
\method{as.data.frame}{device_data}(x, ...)
}
\arguments{
\item{x}{The data returned from \code{\link[=get_data]{get_data()}}.}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of device data
}
\description{
Coerce device data to a data.frame
}
