% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.devices}
\alias{as.data.frame.devices}
\title{Coerce the devices class to a data.frame}
\usage{
\method{as.data.frame}{devices}(x, ...)
}
\arguments{
\item{x}{The devices class returned by \code{get_devices()}}
}
\value{
The data from \code{get_devices()} in data.frame format

A data.frame containing the devices information
}
\description{
Converts the class returned by \code{\link{get_devices}()} to a data.frame.
}
