% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.networks}
\alias{as.data.frame.networks}
\title{Coerce networks information to data.frame}
\usage{
\method{as.data.frame}{networks}(x, ...)
}
\arguments{
\item{x}{The networks class returned by \code{\link[=get_networks]{get_networks()}}}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of networks information
}
\description{
Coerce networks information to data.frame
}
