% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.organizations}
\alias{as.data.frame.organizations}
\title{Coerce organizations information to data.frame}
\usage{
\method{as.data.frame}{organizations}(x, ...)
}
\arguments{
\item{x}{The organizations class returned by \code{\link[=get_organizations]{get_organizations()}}}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of organizations information
}
\description{
Coerce organizations information to data.frame
}
