% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.teams}
\alias{as.data.frame.teams}
\title{Converts teams information to data.frame}
\usage{
\method{as.data.frame}{teams}(x, ...)
}
\arguments{
\item{x}{The teams class returned by \link{\code{get_teams()}}}
}
\value{
A data.frame containing the teams information
}
\description{
Converts teams information to data.frame
}
