% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{authenticate_client}
\alias{authenticate_client}
\title{Authenticate API connection}
\usage{
authenticate_client(verbose = FALSE)
}
\description{
Confirms API authentication
}
\details{
A helper function for \code{client_setup()} to determine that the user-provided
API details yields a proper connection with the API.
}
