% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{build_api_url}
\alias{build_api_url}
\title{Build and return URL}
\usage{
build_api_url(endpoint, qs_params = NULL)
}
\arguments{
\item{endpoint}{A character vector of an endpoint. Can be a full or relative URL}

\item{qs_params}{A list of query string parameters to format and attach to the URL (e.g. list(per_page = 10))}
}
\value{
A character vector containing the final built URL.
}
\description{
Build and return the URL given an endpoint and any query parameters.
}
