% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_data_by_date}
\alias{get_data_by_date}
\title{Get device data by date}
\usage{
get_data_by_date(sn, date, raw = FALSE)
}
\arguments{
\item{sn}{A device serial number}

\item{date}{A date for which to return data. Must be in format "YYYY-MM-DD".}

\item{raw}{(optional) Default FALSE. Returns the raw data. Currently only available to developers and admins.}
}
\value{
The specified device data for the specified date
}
\description{
Get data according to the provided serial number and date.
}
