% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_device_metadata}
\alias{get_device_metadata}
\title{Get device metadata}
\usage{
get_device_metadata(sn)
}
\arguments{
\item{sn}{Device serial number}
}
\value{
The device metadata
}
\description{
Get device metadata
}
