% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_devices}
\alias{get_devices}
\title{Get the user's devices}
\usage{
get_devices(sn = NULL)
}
\arguments{
\item{sn}{A device serial number}
}
\value{
The user's device information
}
\description{
Get the user's devices
}
