% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_logs}
\alias{get_logs}
\title{Get device log data.}
\usage{
get_logs(sn, ...)
}
\arguments{
\item{sn}{Device serial number}

\item{...}{Named arguments}
}
\value{
Data logs
}
\description{
Get device log data.
}
