% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_models}
\alias{get_models}
\title{Get device calibration models}
\usage{
get_models(sn)
}
\arguments{
\item{sn}{Device serial number}
}
\value{
Device calibration models
}
\description{
Get device calibration models
}
