% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_networks}
\alias{get_networks}
\title{Get the user's networks information, within the context of an organization}
\usage{
get_networks(organization_id, network_id = NULL)
}
\arguments{
\item{organization_id}{The unique numeric identifier for the parent org}

\item{network_id}{The unique numeric identifier for the network}
}
\value{
The networks information
}
\description{
Get the user's networks information, within the context of an organization
}
