% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_organizations}
\alias{get_organizations}
\title{Get the user's organizations information.}
\usage{
get_organizations(organization_id = NULL)
}
\arguments{
\item{organization_id}{The unique numeric identifier for the org}
}
\value{
The organizations information
}
\description{
Get the user's organizations information.
}
