% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_teams}
\alias{get_teams}
\title{Get the user's teams information.}
\usage{
get_teams(id = NULL)
}
\arguments{
\item{id}{The unique numeric identifier for the team}
}
\value{
The teams information
}
\description{
Get the user's teams information.
}
