% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{paginate}
\alias{paginate}
\title{Deal with paginated data.}
\usage{
paginate(response_content)
}
\arguments{
\item{request}{Content from a \code{\link{request()}}}
}
\value{
The collated data from every page of the data.
}
\description{
Iterates over and collated all pages of the data.
}
