% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{quantaq_request}
\alias{quantaq_request}
\title{Complete an API request}
\usage{
quantaq_request(endpoint, qs_params = NULL)
}
\arguments{
\item{endpoint}{A character string of the API endpoint to request.}

\item{qs_params}{Query string parameters.}
}
\value{
The parsed JSON from the API, including both data and metadata.
}
\description{
Generate an API request given an endpoint and any relevant query parameters.
}
\examples{
\dontrun{r <- quantaq_request("account")}

# include named arguments for query params, e.g. to limit the response to the first 5 results:
\dontrun{r <- quantaq_request("devices/MOD-PM-00808/data", limit = 5)}

}
