% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{request}
\alias{request}
\title{Complete an API request}
\usage{
request(endpoint, verb = httr::GET, ...)
}
\arguments{
\item{endpoint}{A character string of the API endpoint to request.}

\item{verb}{The httr function corresponding to the HTTP verb. Defaults to GET.}

\item{...}{Query string parameters.}
}
\value{
The parsed JSON from the API, including both data and metadata.
}
\description{
Generate an API request given endpoint, http verb, and any relevant query parameters.
}
\examples{
\dontrun{r <- request("account")}

# include named arguments for query params, e.g. to limit the response to the first 5 results:
\dontrun{r <- request("devices/MOD-PM-00808/data", limit = 5)}

}
