% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{requests}
\alias{requests}
\title{Request that handles pagination}
\usage{
requests(endpoint, ...)
}
\arguments{
\item{endpoint}{A character string of the API endpoint to request.}

\item{...}{Params to be translated to query string.}
}
\value{
Parsed data from the API.
}
\description{
Request that handles pagination
}
