% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_client}
\alias{setup_client}
\title{Setup the API client}
\usage{
setup_client(api_type = "prod", version = "v1", api_key = NULL)
}
\arguments{
\item{api_type}{(Optional) A character string containing the API type -- 'prod' or 'dev' -- which determines the base url. Defaults to 'prod'.}

\item{version}{(Optional) A character string containing the API version you're querying. Defaults to 'v1'.}

\item{api_key}{(Optional) A character string containing your QuantAQ API key. If left NULL, user will be prompted for it during execution.}
}
\value{
No return value. Sets API details as environment variables.
}
\description{
Sets up user details for querying the API
}
\details{
Assigns API key, base url, version, and user agent to environment variables
for access by later API calls, and verifies these settings.
}
