% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{unnest_all}
\alias{unnest_all}
\title{A convenience function to unnest recursively nested data frames.}
\usage{
unnest_all(df)
}
\arguments{
\item{df}{A recursively-nested data frame}
}
\value{
A flattened data frame, one row per observation
}
\description{
A convenience function to unnest recursively nested data frames.
}
