% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.account}
\alias{as.data.frame.account}
\title{Coerce accounts class to data.frame}
\usage{
\method{as.data.frame}{account}(x, ...)
}
\arguments{
\item{x}{The account class returned by \code{\link[=whoami]{whoami()}}}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of the user's account information
}
\description{
Coerce accounts class to data.frame
}
