% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.calibration_models}
\alias{as.data.frame.calibration_models}
\title{Coerce calibration models to data.frame}
\usage{
\method{as.data.frame}{calibration_models}(x, ...)
}
\arguments{
\item{x}{The calibration_models class returned by \code{\link[=get_models]{get_models()}}}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of calibration_models
}
\description{
Coerce calibration models to data.frame
}
