% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.device_data}
\alias{as.data.frame.device_data}
\title{Converts the device data class to a data.frame.}
\usage{
\method{as.data.frame}{device_data}(x, ...)
}
\arguments{
\item{data}{The data returned from \code{\link{get_data()}}.}
}
\value{
The data from \code{get_data()} in data.frame format.
}
\description{
Converts the device data class to a data.frame.
}
