% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.devices}
\alias{as.data.frame.devices}
\title{Coerce devices information to a data.frame}
\usage{
\method{as.data.frame}{devices}(x, ...)
}
\arguments{
\item{x}{The devices class returned by \code{\link[=get_devices]{get_devices()}}}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of device information
}
\description{
Coerce devices information to a data.frame
}
