% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as.data.frame.teams}
\alias{as.data.frame.teams}
\title{Coerce teams information to data.frame}
\usage{
\method{as.data.frame}{teams}(x, ...)
}
\arguments{
\item{x}{The teams class returned by \code{\link[=get_teams]{get_teams()}}}

\item{...}{Placeholder for passing through to as.data.frame.default}
}
\value{
A data.frame of teams information
}
\description{
Coerce teams information to data.frame
}
