% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{build_api_url}
\alias{build_api_url}
\title{Build and return URL}
\usage{
build_api_url(endpoint, ...)
}
\arguments{
\item{endpoint}{A character vector of an endpoint. Can be a full or relative URL}

\item{...}{Query parameters to format and attach to the URL (e.g. per_page)}
}
\value{
A character vector containing the final built URL.
}
\description{
Build and return the URL given an endpoint and any query parameters.
}
