% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_data}
\alias{get_data}
\title{Get device data}
\usage{
get_data(
  sn,
  limit = NULL,
  start = NULL,
  stop = NULL,
  filter = NULL,
  sort = NULL,
  raw = FALSE
)
}
\arguments{
\item{sn}{A device serial number}

\item{limit}{(optional) The number of data points to return}

\item{start}{(optional) The earliest date to retrieve data from. Should be a timestamp string of the form "YYYY-MM-DD HH:MM:SS"}

\item{stop}{(optional) The latest date to retrieve data from. Should be a timestamp string of the form "YYYY-MM-DD HH:MM:SS"}

\item{filter}{(optional) A string providing filter parameters, see below examples and \href{https://docs.quant-aq.com/api#8e14edbf9dee4162a04f729ce022cb4b}{API documentation} for more information.}

\item{sort}{(optional) A data variable upon which to sort, and the sort method (ascending or descending), formatted as "parameter,order", e.g. "timestamp,asc"}

\item{raw}{(optional) Default FALSE. Returns the raw data. Currently only available to developers and admins.}
}
\value{
The specified device data
}
\description{
Get data according to provided serial number and other parameters.
}
\examples{
\dontrun{
  get_data(sn = "MOD-PM-00808", limit = 100)
  get_data(sn = "MOD-PM-00808", start = "2020-01-01 21:54", stop = "2023-03-03 05:47")
  get_data(sn = "MOD-PM-00808", filter = "device_state,eq,ACTIVE")
  get_data(sn = "MOD-PM-00808", sort = "timestamp,asc")
}

}
