% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_devices}
\alias{get_devices}
\title{Get the user's devices}
\usage{
get_devices(sn = NULL, ...)
}
\arguments{
\item{sn}{A device serial number}

\item{...}{Named arguments, passed as query parameters (see: limit, sort)}

\item{limit}{(optional) The number of devices to return}

\item{sort}{(optional) A parameter upon which to sort, and the sort method (ascending or descending), formatted as "parameter,order", e.g. "id,asc"}
}
\value{
The user's device information
}
\description{
Get the user's devices
}
