% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_logs}
\alias{get_logs}
\title{Get device log data.}
\usage{
get_logs(sn, limit = NULL)
}
\arguments{
\item{sn}{Device serial number}

\item{limit}{(optional) The number of logs to return}
}
\value{
Data logs
}
\description{
Get device log data.
}
