% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{request}
\alias{request}
\title{Complete an API request}
\usage{
request(endpoint, verb = httr::GET, query_params = NULL, filter_specs = NULL)
}
\arguments{
\item{endpoint}{A character string of the API endpoint to request.}

\item{verb}{The httr function representing the HTTP verb. Defaults to GET.}

\item{query_params}{(Optional) A named list including any query arguments.}
}
\value{
A request object containing the response from the API.
}
\description{
Generate an API request given endpoint, http verb, and any relevant query parameters.
}
\examples{
\dontrun{r <- request("account")}
\dontrun{r <- request("devices/MOD-PM-00808/data", query_params = list(limit=5))}

}
