% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{requests}
\alias{requests}
\title{Request that handles pagination}
\usage{
requests(endpoint, verb = httr::GET, ...)
}
\arguments{
\item{endpoint}{A character string of the API endpoint to request.}

\item{verb}{The httr function corresponding to the HTTP verb. Defaults to GET.}

\item{...}{Parameters}
}
\value{
Parsed data from the API.
}
\description{
Request that handles pagination
}
